Moving youtube components to a youtube folder
