+++ content_creator_client/src/components/youtube/YoutubeVideoCard.tsx
